Move base64 decoding in to JS library. NFC
