+++ test/code_size/random_printf_wasm.json
@@ -1,6 +1,6 @@
 {
-  "a.html": 12449,
-  "a.html.gz": 6808,
-  "total": 12449,
-  "total_gz": 6808
+  "a.html": 12438,
+  "a.html.gz": 6795,
+  "total": 12438,
+  "total_gz": 6795
 }
